+++ lib/string_translate.dart
@@ -5,3 +5,4 @@ library string_translate;
 
 export 'src/translations_src.dart';
 export 'src/translation_locales.dart';
+export 'src/standart_translations.dart';